fix: #179 - moved default form fields
